+++ src/main/java/com/elmakers/mine/bukkit/magicworlds/populator/MagicBlockPopulator.java
@@ -1,7 +1,5 @@
 package com.elmakers.mine.bukkit.magicworlds.populator;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Random;
 
 import org.bukkit.Chunk;
@@ -9,10 +7,8 @@ import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.configuration.ConfigurationSection;
-import org.bukkit.generator.BlockPopulator;
 
 import com.elmakers.mine.bukkit.magicworlds.MagicWorldsController;
-import com.elmakers.mine.bukkit.magicworlds.populator.builtin.WandChestPopulator;
 
 public abstract class MagicBlockPopulator extends MagicChunkPopulator {
     private int maxY = 255;