+++ src/main/java/com/elmakers/mine/bukkit/magicworlds/populator/MagicChunkPopulator.java
@@ -1,18 +1,9 @@
 package com.elmakers.mine.bukkit.magicworlds.populator;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Random;
-
-import org.bukkit.Chunk;
-import org.bukkit.Material;
-import org.bukkit.World;
-import org.bukkit.block.Block;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.generator.BlockPopulator;
 
 import com.elmakers.mine.bukkit.magicworlds.MagicWorldsController;
-import com.elmakers.mine.bukkit.magicworlds.populator.builtin.WandChestPopulator;
 
 public abstract class MagicChunkPopulator extends BlockPopulator {
 	public static final String BUILTIN_CLASSPATH = "com.elmakers.mine.bukkit.magicworlds.populator.builtin";